Improved image rendering
There was a bug in the new rendering code that caused the top and left sides of the images to be cut off. The fix was to see the pixel offset to half and to make the image box a perfect square that's integer scaled to all of the puzzles.

5x5 puzzles still have some funkiness with their scaling, but other sizes work.